requirements.txt
streamlit
google-generativeai
streamlit-mic-recorder
SpeechRecognition
gTTS
pydub
langdetect